added and exported getAllTodos function in dbHelpers.js
added a getAll route and renamed the add route in todoRoutes.js, along with a hardcoded limit value for testing purposes

script.js: added a load_by constant and created new interface generation functions called makeGrid and insertIntoGrid. their aim is to be more flexible than the original makeGroup/makeTodo functions. those functions were great at fulfilling the mockup style, but rely on loading groups specifically to correctly format the todos. This methodology will not work as cleanly if the user wants to load todos without group information. For now the old functions will be left in, as they may be useful for design in the future.

app.ejs: added test code for the makeGrid(function)

homepage.ejs: added test code for the new getAll route