committing piecewise work done during break: contains various progress on several features, including structure and JSON information flow changes
I took a break from the development of this project in order to step back and rethink my approach and structure. So far I've realized that the current implementation of the classes in the models folder is redundant as functionally the objecs are turned back into plain JSON objects during use. It may be possible to convert them client-side in order to do operations on them, but I haven't weighed the usefulness of the models as a whole.

for more detailed code changes, see the diff and comments below.